Share the switch layers
